src/litecow/pyproject.toml
[tool.poetry]
name = "litecow"
version = "0.1.0"
description = ""
authors = ["Striveworks"]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.33.2"
protobuf = "*"
numpy = "^1.17.4"
onnxruntime = {version = "^1.7.0", optional = true }
onnxruntime-gpu = {version = "^1.7.0", optional = true }
boto3 = {version = "^1.11.14", optional = true }

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.34.0"

[tool.poetry.scripts]
litecow-service = 'litecow.server:main'

[tool.poetry.extras]
server-gpu = ["onnxruntime-gpu","boto3"]
server-cpu = ["onnxruntime","boto3"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
